Match "i10000," against an integer intead of string in test
All Cencode/BBE development up to this point was under Cygwin and this
test never failed there.  I saw it first fail under Linux. If I patch
Test::Differences under Cygwin to use Data::Dumper::XS then the
original test fails there as well (as it should have) so there might be
an issue with the pure Perl version of Data::Dumper.