UTF8 and BYTES strings now encode with a trailing ","
This is a last-minute change to the encoding specification. Similarly
to netstrings we now require a "," at the end of every utf8/byte
string.  The increase in encoding size is a trade-off I'm willing to
make for two reasons: It adds a small check to ensure that the end of
the string is where it is expected, but more importantly it improves
readability of arrays/lists and hashes/dictionaries. That makes
implementing and testing bifcode a good chunk easier.